requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.1.1=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
albucore=0.0.20=pypi_0
albumentations=1.4.21=pypi_0
altair=5.4.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.6.2.post1=pypi_0
argparse=1.4.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.7=py311hfdbb021_0
decorator=5.1.1=pyhd8ed1ab_0
diffusers=0.31.0=pypi_0
docker-pycreds=0.4.0=pypi_0
eval-type-backport=0.2.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
fonttools=4.54.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.2.0=pypi_0
ftfy=6.3.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
h5py=3.12.1=pypi_0
idna=3.10=pypi_0
imageio=2.36.0=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.5.0=pyha770c72_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.29.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
json-repair=0.30.0=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
kaggle=1.6.17=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=pypi_0
kornia=0.7.4=pypi_0
kornia-rs=0.1.7=pypi_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.43=h712a8e2_2
libedit=3.1.20191231=he28a2e2_2
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgomp=14.2.0=h77fa898_1
libnsl=2.0.1=hd590300_0
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.47.0=hadc24fc_1
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
lightning-utilities=0.11.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
narwhals=1.13.3=pypi_0
ncurses=6.5=he02047a_1
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.3.1=pyh8b19718_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
propcache=0.2.0=pypi_0
psutil=6.1.0=py311h9ecbd09_0
ptyprocess=0.7.0=pyhd3deb0d_0
pudb=2024.1.3=pypi_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=18.0.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.2.0=pypi_0
python=3.11.10=hc5c86c4_3_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-slugify=8.0.4=pypi_0
python_abi=3.11=5_cp311
pytorch-lightning=2.4.0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py311h7deb3e3_3
readline=8.2=h8228510_1
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
sentry-sdk=2.17.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=75.3.0=pyhd8ed1ab_0
simsimd=6.0.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.47.0=h9eae976_1
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.40.1=pypi_0
stringzilla=3.10.10=pypi_0
sympy=1.13.1=pypi_0
taming-transformers=0.0.1=pypi_0
tenacity=9.0.0=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.20.1=pypi_0
toml=0.10.2=pypi_0
torch-fidelity=0.3.0=pypi_0
torchaudio=2.4.0+cu121=pypi_0
torchmetrics=1.5.2=pypi_0
tornado=6.4.1=py311h9ecbd09_1
tqdm=4.66.6=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.46.1=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
urwid=2.6.16=pypi_0
urwid-readline=0.15.1=pypi_0
validators=0.34.0=pypi_0
wandb=0.18.5=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yarl=1.17.1=pypi_0
zeromq=4.3.5=h3b0a872_6
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.3.1=hb9d3cd8_2
